+++ src/pandemia/components/travel_model/safir_travel_model.py
@@ -5,7 +5,7 @@ import numpy as np
 import copy
 from joblib import Parallel, delayed
 
-from ctypes import c_void_p, c_double, c_int32, cdll
+from ctypes import c_void_p, c_double, c_int64, cdll
 
 from ..travel_model import TravelModel
 log = logging.getLogger("safir_travel_model")
@@ -57,7 +57,7 @@ class SafirTravelModel(TravelModel):
 
         self.travel_multiplier = config['travel_transmission_multiplier']
         self.interpolation     = config['interpolation']
-        self.beta              = np.array(config['beta'], dtype=float)
+        self.beta              = np.array(config['beta'], dtype=np.float64)
 
         self.baseline_agents_travelling_matrix = vector_world.travel_matrix
         for r in range(self.number_of_regions):
@@ -68,7 +68,7 @@ class SafirTravelModel(TravelModel):
 
         vector_region.current_border_closure_multiplier = 1.0
         vector_region.current_region = np.full((vector_region.number_of_agents),
-                                                vector_region.id, dtype=np.int32)
+                                                vector_region.id, dtype=np.int64)
 
     def initial_conditions(self, sim):
         """Initial regional mixing"""
@@ -87,9 +87,9 @@ class SafirTravelModel(TravelModel):
 
         for vector_region in vector_region_batch:
             self.determine_travellers(
-                c_int32(vector_region.number_of_agents),
-                c_int32(vector_region.id),
-                c_int32(self.number_of_regions),
+                c_int64(vector_region.number_of_agents),
+                c_int64(vector_region.id),
+                c_int64(self.number_of_regions),
                 c_void_p(vector_region.current_disease.ctypes.data),
                 c_void_p(vector_region.current_strain.ctypes.data),
                 c_void_p(vector_region.current_region.ctypes.data),
@@ -97,8 +97,8 @@ class SafirTravelModel(TravelModel):
                 c_void_p(vector_region.random_state.ctypes.data)
             )
             self.transmission_out(
-                c_int32(vector_region.number_of_agents),
-                c_int32(vector_region.id),
+                c_int64(vector_region.number_of_agents),
+                c_int64(vector_region.id),
                 c_void_p(self.beta.ctypes.data),
                 c_double(facemask_transmission_multiplier),
                 c_double(self.travel_multiplier),
@@ -117,8 +117,8 @@ class SafirTravelModel(TravelModel):
 
         for vector_region in vector_region_batch:
             self.transmission_in(
-                c_int32(vector_region.number_of_agents),
-                c_int32(vector_region.id),
+                c_int64(vector_region.number_of_agents),
+                c_int64(vector_region.id),
                 c_void_p(vector_region.current_facemask.ctypes.data),
                 c_void_p(vector_region.current_region.ctypes.data),
                 c_double(facemask_transmission_multiplier),
@@ -136,7 +136,7 @@ class SafirTravelModel(TravelModel):
             vector_region.start_t[indexes] = time
             vector_region.current_strain[indexes] = 0
             vector_region.infection_event =\
-                np.full(vector_region.number_of_agents, -1, dtype=np.int32)
+                np.full(vector_region.number_of_agents, -1, dtype=np.int64)
 
     def dynamics(self, sim, day, ticks_in_day,
                  facemask_transmission_multiplier,
@@ -145,15 +145,15 @@ class SafirTravelModel(TravelModel):
         """Changes to regional mixing model"""
 
         agents_travelling_matrix =\
-            np.zeros((self.number_of_regions, self.number_of_regions), dtype=np.int32)
+            np.zeros((self.number_of_regions, self.number_of_regions), dtype=np.int64)
 
         # Reset record of who is travelling abroad and apply border closure if necessary
         for vector_region in sim.vector_regions:
             vector_region.current_region = np.full((vector_region.number_of_agents),
-                                                   vector_region.id, dtype=np.int32)
+                                                   vector_region.id, dtype=np.int64)
             self.close_borders(
-                c_int32(self.number_of_regions),
-                c_int32(vector_region.id),
+                c_int64(self.number_of_regions),
+                c_int64(vector_region.id),
                 c_double(self.scale_factor),
                 c_double(vector_region.current_border_closure_multiplier),
                 c_void_p(agents_travelling_matrix.ctypes.data),
@@ -162,9 +162,9 @@ class SafirTravelModel(TravelModel):
 
         # Update record of who is travelling abroad and calculate transmission force for each region
         sum_f_by_strain =\
-            np.zeros((self.number_of_regions * self.number_of_strains), dtype=float)
+            np.zeros((self.number_of_regions * self.number_of_strains), dtype=np.float64)
         transmission_force =\
-            np.ones((self.number_of_regions), dtype=float)
+            np.ones((self.number_of_regions), dtype=np.float64)
         if enable_parallel:
             Parallel(n_jobs=num_jobs, backend="threading",
                      verbose=0)(delayed(self._out)(vector_region_batch, agents_travelling_matrix,
@@ -189,8 +189,8 @@ class SafirTravelModel(TravelModel):
         """Artificially inflate number of travellers for testing purposes or otherwise"""
 
         ids_to_population_sizes =\
-            {vr.id: np.int32(vr.number_of_agents * (1 / self.scale_factor))\
-                             for vr in vector_world.vector_regions}
+            {vr.id: np.float64(vr.number_of_agents * (1 / self.scale_factor))\
+                    for vr in vector_world.vector_regions}
 
         # Adjust matrix using interpolation parameter
         row_sums = baseline_agents_travelling_matrix.sum(axis=1)
